requirements.txt
djangorestframework==3.12.2
Django==3.1.7
django-rest-auth==0.9.5
django-filter==2.4.0
django-cors-headers==3.7.0
Pygments==2.8.1
Markdown==3.3.4
coreapi==2.3.3
psycopg2==2.8.6
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
